# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Saray Property Booking System",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "URL"
        }
    ],
    "paths": {
...
        "/api/payment-plans/{id}": {
            "get": {
                "tags": [
                    "PaymentPlans"
                ],
                "summary": "Retrieve a specific payment plan",
                "operationId": "showPaymentPlan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment plan to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment plan retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment_plan": {
                                            "$ref": "#/components/schemas/PaymentPlan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — user lacks permission to view this plan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found — no plan exists with the given ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\PaymentPlan] 123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PaymentPlans"
                ],
                "summary": "Delete a payment plan",
                "description": "Delete a payment plan and its installments.",
                "operationId": "23bb54c79dc39bfd4f313b7b4f8c29e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment plan to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payment plan deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden — user lacks permission to delete this plan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found — payment plan does not exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\PaymentPlan] 123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict — cannot delete a plan applied to existing bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Cannot delete a payment plan that is applied on existing bookings."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentPlan": {
                "title": "Payment Plan",
                "required": [
                    "unit_id",
                    "name",
                    "dld_fee_percentage",
                    "admin_fee",
                    "EOI",
                    "booking_percentage",
                    "handover_percentage",
                    "construction_percentage"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "unit_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "60/40"
                    },
                    "dld_fee_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 65000
                    },
                    "admin_fee": {
                        "type": "number",
                        "format": "float",
                        "example": 4000
                    },
                    "EOI": {
                        "type": "number",
                        "format": "float",
                        "example": 100000
                    },
                    "booking_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "handover_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 40
                    },
                    "construction_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 40
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-01-02T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}